+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bug #216: Extract `reinitializeApplication()` in `StatelessApplication` as a protected request-preparation hook (@terabytesoftw)
 - Bug #217: Extract `resetRequestState()` in `StatelessApplication` as a protected request-preparation hook (@terabytesoftw)
 - Bug #218: Rename `ApplicationRestTest` to `ApplicationHookTest` to reflect its focus on testing lifecycle hooks (@terabytesoftw)
+- Bug #219: Extract `prepareErrorHandler()` in `StatelessApplication` as a protected request-preparation hook (@terabytesoftw)
 
 ## 0.1.5 January 28, 2026